apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ops-copilot"
version = "0.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.128.5",
    "uvicorn>=0.30",
    "SQLAlchemy>=2.0",
    "pydantic-settings>=2.0",
    "ops-copilot-db",
    "ops-copilot-agent-runtime",
    "pytest>=8",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests/unit", "tests/integration"]
python_files = ["test_*.py", "*_test.py"]
pythonpath = ["src"]
markers = [
    "integration: integration tests requiring real services",
]
